GO-7320 Remove subscription engine, leaving the API surface unimplemented
Blank-slate starting point for reimplementing object subscriptions from
scratch. The engine (sorted/simple/ids/collection/group subscriptions,
entry cache, dependency service, event assembly, change batching) is
deleted; the package keeps only what consumers compile against:

- Service interface, SubscribeRequest/Response, SubscribeGroupsRequest,
  CollectionService and New() returning a stub: query methods return
  ErrNotImplemented, teardown methods are no-ops
- EventMatcher (consumer-side queue utility)
- test fixtures and generated mocks used by other packages

crossspacesub and objectsubscription stay: they are consumers/adapters
above the Service interface, not the engine.

The contract the reimplementation must fulfill is documented in
docs/Subscriptions.md, derived from real usage in anytype-ts and
anytype-heart internals.
